# pkg/trace/agent: add stream decoder