This change adds a stream decoder which decodes traces and sends them for processing one at a time, instead of decoding everything and then iterating through everything once more to process.

Benchmark results are:
```
name                     old time/op    new time/op     delta
Throughput/39KB/10-4        584µs ± 0%      503µs ± 0%  -13.92%
Throughput/1.1MB/100-4     11.5ms ± 0%      8.7ms ± 0%  -24.06%
Throughput/1.1MB/1951-4    18.0ms ± 0%     15.1ms ± 0%  -16.09%
Throughput/2.2MB/4161-4    40.2ms ± 0%     31.8ms ± 0%  -20.98%
Throughput/6.6MB/500-4     73.5ms ± 0%     57.3ms ± 0%  -22.06%

name                     old speed      new speed       delta
Throughput/39KB/10-4     70.6MB/s ± 0%   82.0MB/s ± 0%  +16.17%
Throughput/1.1MB/100-4   97.4MB/s ± 0%  128.3MB/s ± 0%  +31.69%
Throughput/1.1MB/1951-4  63.1MB/s ± 0%   75.2MB/s ± 0%  +19.17%
Throughput/2.2MB/4161-4  59.8MB/s ± 0%   75.7MB/s ± 0%  +26.56%
Throughput/6.6MB/500-4   96.9MB/s ± 0%  124.3MB/s ± 0%  +28.32%

name                     old alloc/op   new alloc/op    delta
Throughput/39KB/10-4        234kB ± 0%      233kB ± 0%   -0.49%
Throughput/1.1MB/100-4     5.60MB ± 0%     5.51MB ± 0%   -1.69%
Throughput/1.1MB/1951-4    7.30MB ± 0%     7.18MB ± 0%   -1.67%
Throughput/2.2MB/4161-4    15.4MB ± 0%     15.1MB ± 0%   -1.57%
Throughput/6.6MB/500-4     37.6MB ± 0%     37.0MB ± 0%   -1.75%

name                     old allocs/op  new allocs/op   delta
Throughput/39KB/10-4        3.37k ± 0%      3.24k ± 0%   -3.91%
Throughput/1.1MB/100-4      79.7k ± 0%      75.5k ± 0%   -5.21%
Throughput/1.1MB/1951-4      145k ± 0%       138k ± 0%   -5.16%
Throughput/2.2MB/4161-4      307k ± 0%       292k ± 0%   -5.05%
Throughput/6.6MB/500-4       524k ± 0%       497k ± 0%   -5.23%
````