@gbbr 2019-03-28
Next week is release testing week for 6.11 and I'll be monitoring it extensively.